# frontend/swagger.json
{"openapi":"3.0.0","paths":{"/user":{"get":{"operationId":"UserController_getAll","parameters":[{"name":"search","required":false,"in":"query","description":"Search by name or characterId","schema":{"type":"string"}}],"responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"summary":"Get users (all or filtered)","tags":["Users"]}},"/user/me":{"get":{"operationId":"UserController_getMe","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get full profile of the authenticated user","tags":["Users"]}},"/user/search":{"get":{"operationId":"UserController_searchUsers","parameters":[{"name":"q","required":true,"in":"query","description":"Search query","schema":{"example":"eve","type":"string"}}],"responses":{"200":{"description":""}},"summary":"Search users","tags":["Users"]}},"/user/{id}":{"get":{"operationId":"UserController_getOne","parameters":[{"name":"id","required":true,"in":"path","description":"User ID (cuid)","schema":{"type":"string"}}],"responses":{"200":{"description":"User found","content":{"application/json":{"schema":{"type":"object"}}}}},"summary":"Get user by ID","tags":["Users"]}},"/user/transactions":{"get":{"operationId":"UserController_getMyTransactions","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get transactions of the authenticated user","tags":["Users"]}},"/user/set-referral":{"post":{"operationId":"UserController_setReferral","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetReferralDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Users"]}},"/user/become-executor":{"post":{"operationId":"UserController_becomeExecutor","parameters":[],"responses":{"200":{"description":"Role updated to EXECUTOR"}},"security":[{"bearer":[]}],"summary":"Become executor","tags":["Users"]}},"/transaction/deposit":{"post":{"operationId":"TransactionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDepositDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create a deposit request to top up your balance","tags":["Transactions"]}},"/transaction":{"get":{"operationId":"TransactionController_getAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"example":20,"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get your transaction history","tags":["Transactions"]}},"/transaction/withdraw":{"post":{"operationId":"TransactionController_requestWithdraw","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"amount":{"type":"number","example":1000000000}},"required":["amount"]}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Request a withdrawal from your internal balance","tags":["Transactions"]}},"/transaction/withdraw/{id}":{"delete":{"operationId":"TransactionController_cancelWithdraw","parameters":[{"name":"id","required":true,"in":"path","description":"Transaction ID of the withdrawal","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Cancel a pending withdrawal request","tags":["Transactions"]}},"/transaction/corporate":{"get":{"operationId":"TransactionController_getBalance","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get current corporation wallet balance","tags":["Transactions"]}},"/order/types":{"get":{"operationId":"OrderController_getTypes","parameters":[],"responses":{"200":{"description":""}},"summary":"Get available order types (enum values)","tags":["Orders"]}},"/order":{"get":{"operationId":"OrderController_getAll","parameters":[{"name":"userId","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"type","required":false,"in":"query","schema":{"enum":["KILL_TARGET","SCAN_WORMHOLE","SCOUT_SYSTEM","LOGISTICS","ESCORT","STRUCTURE_WORK","CHARACTER_INFO","ROUTE_PLANNING","COUNTER_INTEL","EVENT_FARMING","PVP_ASSIST","OTHER"],"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"enum":["ACTIVE","TAKEN","DONE","CANCELED"],"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get orders (optionally filtered)","tags":["Orders"]},"post":{"operationId":"OrderController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create a new order","tags":["Orders"]}},"/order/promoted":{"get":{"operationId":"OrderController_getPromotedOrders","parameters":[],"responses":{"200":{"description":""}},"summary":"Get promoted orders (sorted by oldest + most expensive)","tags":["Orders"]}},"/order/{id}":{"get":{"operationId":"OrderController_getOne","parameters":[{"name":"id","required":true,"in":"path","description":"Order ID","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get order by ID","tags":["Orders"]},"delete":{"operationId":"OrderController_delete","parameters":[{"name":"id","required":true,"in":"path","description":"Order ID","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete order (creator only)","tags":["Orders"]}},"/order/{id}/take":{"post":{"operationId":"OrderController_take","parameters":[{"name":"id","required":true,"in":"path","description":"Order ID","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Take order as executor","tags":["Orders"]}},"/order/{id}/status":{"patch":{"operationId":"OrderController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","description":"Order ID","schema":{"type":"string"}},{"name":"status","required":true,"in":"query","schema":{"enum":["ACTIVE","TAKEN","DONE","CANCELED"],"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update order status (DONE or CANCELED)","tags":["Orders"]}},"/order/{id}/message":{"post":{"operationId":"OrderController_sendMessage","parameters":[{"name":"id","required":true,"in":"path","description":"Order ID","schema":{"type":"string"}},{"name":"text","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Send message to order chat","tags":["Orders"]}},"/order/{id}/messages":{"get":{"operationId":"OrderController_getMessages","parameters":[{"name":"id","required":true,"in":"path","description":"Order ID","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get messages for order","tags":["Orders"]}},"/order/{id}/complain":{"post":{"operationId":"OrderController_complain","parameters":[{"name":"id","required":true,"in":"path","description":"Order ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitComplaintDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Submit a complaint for an order","tags":["Orders"]}},"/order/{id}/review":{"post":{"operationId":"ReviewController_create","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Leave a review for a completed order","tags":["Reviews"]}},"/order/{id}/review/user/{userId}":{"get":{"operationId":"ReviewController_getUserReviews","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get reviews for a specific user","tags":["Reviews"]}},"/auth/eve":{"get":{"operationId":"AuthController_redirectToEve","parameters":[],"responses":{"200":{"description":""}},"summary":"Редирект на EVE Online SSO","tags":["Auth"]}},"/auth/eve/callback":{"get":{"operationId":"AuthController_eveGetCallback","parameters":[{"name":"code","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Auth"]},"post":{"operationId":"AuthController_eveCallback","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"string","example":"abc123..."}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","example":"jwt-token-here"}}}}}}},"summary":"Получить JWT по коду от EVE","tags":["Auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Получить текущего пользователя по JWT","tags":["Auth"]}},"/auth/discord/callback":{"get":{"operationId":"AuthController_discordCallback","parameters":[{"name":"code","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/link":{"post":{"operationId":"AuthController_linkDiscord","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","example":"235822777678954496"}},"required":["id"]}}}},"responses":{"200":{"description":"Discord account linked"},"400":{"description":"Missing or invalid Discord ID"}},"security":[{"bearer":[]}],"summary":"Link Discord account to current user","tags":["Auth"]}},"/system":{"get":{"operationId":"SystemController_getAll","parameters":[],"responses":{"200":{"description":""}},"summary":"Get all systems","tags":["Systems"]}},"/system/search":{"get":{"operationId":"SystemController_search","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Search systems by name (min 3 chars)","tags":["Systems"]}},"/system/info/{systemId}":{"get":{"operationId":"SystemController_getSystemInfo","parameters":[{"name":"systemId","required":true,"in":"path","schema":{"example":30000142,"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Get EVE system info from ESI (auth required)","tags":["Systems"]}}},"info":{"title":"EVE Board API","description":"API для заказов и исполнителей в EVE Online","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"/api"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SetReferralDto":{"type":"object","properties":{}},"CreateDepositDto":{"type":"object","properties":{"amount":{"type":"number","example":10000000},"reference":{"type":"string","example":"user123"}},"required":["amount"]},"CreateOrderDto":{"type":"object","properties":{}},"SubmitComplaintDto":{"type":"object","properties":{}},"CreateReviewDto":{"type":"object","properties":{"rating":{"type":"number","example":5,"description":"Rating from 1 to 5"},"text":{"type":"string","example":"Completed the task quickly and accurately","description":"Optional review comment"}},"required":["rating"]}}}}